pyproject.toml
[tool.poetry]
name = "inforce-test-task"
version = "0.1.0"
description = ""
authors = ["Владислав <vlad.slinchuk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
psycopg2 = "^2.9.3"
drf-yasg = "^1.21.3"
django-cors-headers = "^3.13.0"
flake8 = "^5.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
